% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{AMS}
\alias{AMS}
\title{Generate Approximate Median Significance (AMS) function for specified threshold}
\usage{
AMS(f, valid_set, valid_y, valid_weights)
}
\arguments{
\item{f}{Weighted logistic regression model (output of caret::train).}

\item{valid_set}{Validation data set for model `f`.}

\item{valid_y}{Binary labels of validation data (0/1).}

\item{valid_weights}{Weights of validation data.}
}
\value{
Function with one input `theta`, that returns AMS of validation set using model `f` with threshold `theta`.
}
\description{
Generates an AMS function with threshold (`theta`) of a logistic regression model as input.
 Useful for tuning threshold as a second stage to fitting logistic regression model.
}
