% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{AMS_base}
\alias{AMS_base}
\title{Approximate Median Significance (AMS) Base Function}
\usage{
AMS_base(s, b, b_reg = 10)
}
\arguments{
\item{s}{Sum of weights corresponding to true positive data points.}

\item{b}{Sum of weights corresponding to false positive data points.}

\item{b_reg}{Regularization parameter.}
}
\value{
Approximate median significance.
}
\description{
Approximate Median Significance (AMS) Base Function
}
