% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{AMS_measure}
\alias{AMS_measure}
\title{Generate AMS measure to be used in mlr training}
\usage{
AMS_measure()
}
\value{
AMS measure that can be used as a measure in functions in the `mlr` package, e.g. `crossval`
}
\description{
Generate AMS measure to be used in mlr training
}
