% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{AMS_weighted}
\alias{AMS_weighted}
\title{Approximate Median Significance (AMS) function using weights of data points}
\usage{
AMS_weighted(truth, response, weights)
}
\arguments{
\item{truth}{The true labels (0/1) of the data points.}

\item{response}{The predicted labels (0/1) of the data points.}

\item{weights}{The weights of the data points (can be un-normalized)}
}
\value{
AMS.
}
\description{
Approximate Median Significance (AMS) function using weights of data points
}
