% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{reweight}
\alias{reweight}
\title{Re-weighting}
\usage{
reweight(weights, labels, N_s, N_b)
}
\arguments{
\item{weights}{Input weights (unnormalized).}

\item{labels}{Binary labels of the data points corresponding to the input weights (labels should be 0/1).}

\item{N_s}{Desired sum of weights corresponding to data points with labels equal to 1.}

\item{N_b}{Desired sum of weights corresponding to data points with labels equal to 1.}
}
\value{
Normalized weights
}
\description{
Re-weighting to preserve the sum of weights in each binary class
}
