% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{threshold_CV}
\alias{threshold_CV}
\title{Cross-validation function for tuning threshold in logistic regression.}
\usage{
threshold_CV(df, label, weights, theta_0, theta_1, k = 5, n = 50)
}
\arguments{
\item{df}{Data-frame to perform cross-validation on.}

\item{label}{Binary labels of data points in `df` (0/1).}

\item{weights}{Weights of data points in `df`.}

\item{theta_0}{Lower bound of threshold parameter.}

\item{theta_1}{Upper bound of threshold parameter.}

\item{k}{Number of cross-validation sets.}

\item{n}{Number of values of threshold to check.}
}
\value{
`max_theta` is the threshold value that maximizes the AMS and `max_AMS` is the maximum average AMS found across cross-validation sets.
}
\description{
Function that performs k-fold cross-validation in order to tune the threshold parameter in logistic regression.
 Fits a weighted logistic regression using randomized training/validation split, then find the the threshold parameter that
 maximises the approximate median significance (AMS).
}
